+++ library/src/methods/getDescription/index.ts
@@ -1,0 +1,1 @@
+export * from './getDescription.ts';